Add GCObject/GCUtils for GCHandle interop, update version
- Added GCObject<T> and GCUtils for safe GCHandle pointer interop.
- Introduced NotNullWhenAttribute for .NET Standard 2.0 support.
- Updated MurmurHash3 finalization to use int length for XOR.
- Bumped package version to 2.2.8.

+++ Hexa.NET.Utilities/Hashing/MurmurHash3.cs
@@ -144,8 +144,8 @@
                 }
 
                 // finalization
-                h1 ^= (ulong)length;
-                h2 ^= (ulong)length;
+                h1 ^= length;
+                h2 ^= length;
 
                 h1 += h2;
                 h2 += h1;
